+++ test/dom_tools_test.dart
@@ -1,4 +1,6 @@
 @TestOn('browser')
+library;
+
 import 'dart:html';
 
 import 'package:dom_tools/dom_tools.dart';